requirements.txt
aiohttp>=3.6.2
async-timeout>=3.0.1
attrs>=20.1.0
certifi>=2020.6.20
chardet>=3.0.4
discord.py>=1.4.1
idna>=2.10
multidict>=4.7.6
requests>=2.24.0
urllib3>=1.25.10
yarl>=1.5.1
